make fake timer export named
