+++ packages/jest-environment-node/src/index.js
@@ -13,7 +13,7 @@ import type {Global} from 'types/Global';
 import type {ModuleMocker} from 'jest-mock';
 
 import vm from 'vm';
-import FakeTimers from '@jest/fake-timers';
+import {JestFakeTimers as FakeTimers} from '@jest/fake-timers';
 import {installCommonGlobals} from 'jest-util';
 import mock from 'jest-mock';
